requirements.txt
anthropic>=0.40.0
feedparser>=6.0.10
requests>=2.31.0
beautifulsoup4>=4.12.0
gspread>=6.0.0
google-auth>=2.30.0
openpyxl>=3.1.0
flask>=3.0.0
flask-cors>=4.0.0
